refactor: scanner and settings package
