+++ app/src/main/java/com/dicoding/motour/presentation/setting/AboutActivity.kt
@@ -1,23 +1,21 @@
-package com.dicoding.motour
+package com.dicoding.motour.presentation.setting
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.dicoding.motour.databinding.ActivityFaqBinding
+import com.dicoding.motour.R
+import com.dicoding.motour.databinding.ActivityAboutBinding
 
-class FaqActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityFaqBinding
+class AboutActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityAboutBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         enableEdgeToEdge()
 
         super.onCreate(savedInstanceState)
-        binding = ActivityFaqBinding.inflate(layoutInflater)
+        binding = ActivityAboutBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         setupEdgeToEdge()
@@ -35,13 +33,6 @@ class FaqActivity : AppCompatActivity() {
     private fun setupClickListener() {
         binding.btnBack.setOnClickListener {
             onBackPressedDispatcher.onBackPressed()
-        }
-        binding.tvFaqOneHyperlink.setOnClickListener {
-            val intent = Intent(
-                Intent.ACTION_VIEW,
-                Uri.parse("https://support.google.com/android/answer/9431959")
-            )
-            startActivity(intent)
         }
     }
 }
