+++ app/src/main/java/com/dicoding/motour/presentation/setting/FaqActivity.kt
@@ -1,20 +1,24 @@
-package com.dicoding.motour
+package com.dicoding.motour.presentation.setting
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.dicoding.motour.databinding.ActivityAboutBinding
+import com.dicoding.motour.R
+import com.dicoding.motour.databinding.ActivityFaqBinding
 
-class AboutActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityAboutBinding
+class FaqActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityFaqBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         enableEdgeToEdge()
 
         super.onCreate(savedInstanceState)
-        binding = ActivityAboutBinding.inflate(layoutInflater)
+        binding = ActivityFaqBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         setupEdgeToEdge()
@@ -32,6 +36,13 @@ class AboutActivity : AppCompatActivity() {
     private fun setupClickListener() {
         binding.btnBack.setOnClickListener {
             onBackPressedDispatcher.onBackPressed()
+        }
+        binding.tvFaqOneHyperlink.setOnClickListener {
+            val intent = Intent(
+                Intent.ACTION_VIEW,
+                Uri.parse("https://support.google.com/android/answer/9431959")
+            )
+            startActivity(intent)
         }
     }
 }
